+++ assignment-impl/src/main/webapp/WEB-INF/web.xml
@@ -6,7 +6,7 @@
         version="5.0">
     <servlet>
         <servlet-name>CustomerServlet</servlet-name>
-        <servlet-class>com.epam.rd.davydova.assignment.servlets.CustomerServlet</servlet-class>
+        <servlet-class>com.epam.rd.davydova.assignment.domain.servlets.CustomerServlet</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>CustomerServlet</servlet-name>
@@ -14,7 +14,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>SupplierServlet</servlet-name>
-        <servlet-class>com.epam.rd.davydova.assignment.servlets.SupplierServlet</servlet-class>
+        <servlet-class>com.epam.rd.davydova.assignment.domain.servlets.SupplierServlet</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>SupplierServlet</servlet-name>
@@ -22,7 +22,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>ProductServlet</servlet-name>
-        <servlet-class>com.epam.rd.davydova.assignment.servlets.ProductServlet</servlet-class>
+        <servlet-class>com.epam.rd.davydova.assignment.domain.servlets.ProductServlet</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>ProductServlet</servlet-name>
@@ -30,7 +30,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>OrderServlet</servlet-name>
-        <servlet-class>com.epam.rd.davydova.assignment.servlets.OrderServlet</servlet-class>
+        <servlet-class>com.epam.rd.davydova.assignment.domain.servlets.OrderServlet</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>OrderServlet</servlet-name>
